Fix Rounding of Hours

+++ R/tide-height.R
@@ -54,7 +54,7 @@ hours_year <- function(datetime) {
 
   startdatetime <- ISOdate(year, 1, 1, 0, tz = "UTC")
   hours <- difftime(datetime, startdatetime, units = 'hours')
-  hours %<>% as.integer()
+  hours %<>% as.numeric()
   hours
 }
